requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==40.0.1
daphne==4.0.0
Django==4.2
django-cors-headers==3.14.0
django-crontab==0.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
marshmallow==3.19.0
packaging==23.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dotenv==1.0.0
pytz==2023.3
rest-framework-simplejwt==0.0.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
zope.interface==6.0
